requirements.txt
requests>=2.28.0
pyyaml>=6.0
base64
urllib3>=1.26.0
flask>=2.0.0
jinja2>=3.0.0
click>=8.0.0
colorama>=0.4.0
tqdm>=4.64.0